[JAVAWEB-MOVIEDELIVERY] Cadastro de midia com N - > N adicionado

+++ MovieDelivery-ejb/src/java/br/com/moviedelivery/service/ICategoriaService.java
@@ -9,6 +9,6 @@ public interface ICategoriaService {
     
     public String salvar(Categoria categoria);
     
-    public String remover(Integer id);
+    public String remover(Short id);
     
 }
